+++ src/ReplicateFilesystemAdapter.php
@@ -7,8 +7,14 @@ namespace Ajgl\Flysystem\Replicate;
 use League\Flysystem\ChecksumProvider;
 use League\Flysystem\FilesystemAdapter;
 use League\Flysystem\UrlGeneration\PublicUrlGenerator;
+use League\Flysystem\UrlGeneration\TemporaryUrlGenerator;
 
-if (interface_exists(ChecksumProvider::class)) {
+if (interface_exists(TemporaryUrlGenerator::class)) {
+    final class ReplicateFilesystemAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvider, TemporaryUrlGenerator
+    {
+        use ReplicateFilesystemAdapterTrait;
+    }
+} elseif (interface_exists(ChecksumProvider::class)) {
     final class ReplicateFilesystemAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvider
     {
         use ReplicateFilesystemAdapterTrait;